+++ tests/Support/TestRouteName.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Router\Tests\Support;
+
+enum TestRouteName: string
+{
+    case Home = 'home';
+    case About = 'about';
+}